파일 삭제 코드 수정

+++ HowsServer/src/main/java/com/hows/common/GCSFile.java
@@ -1,18 +1,23 @@
 package com.hows.common;
 
-import com.google.cloud.storage.Blob;
-import com.google.cloud.storage.BlobId;
-import com.google.cloud.storage.BlobInfo;
-import com.google.cloud.storage.Storage;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.google.cloud.storage.Blob;
+import com.google.cloud.storage.BlobId;
+import com.google.cloud.storage.BlobInfo;
+import com.google.cloud.storage.Storage;
+
 @Component
 public class GCSFile {
 
-    @Autowired
+	private static final Logger logger = LoggerFactory.getLogger(GCSFile.class);
+    
+	@Autowired
     private Storage storage;
 
     @Value("${gcp.bucket}")
@@ -23,12 +28,12 @@ public class GCSFile {
     }
 
     /** GCS 파일 서버에 파일 및 이미지 업로드 **/
-    public String addFileGcs (MultipartFile file, String sysName, String type) {
+    public String addFileGcs (MultipartFile file, String sysName, String code) {
         try{
-            BlobId blobId = BlobId.of(bucket, sysName);
+            BlobId blobId = BlobId.of(bucket, code + "/" + sysName);
             BlobInfo blobInfo = BlobInfo.newBuilder(blobId).build();
             Blob blob = storage.create(blobInfo, file.getBytes());
-            if(blob != null) return "https://storage.cloud.google.com/" + bucket + "/" + type + "/" + sysName;
+            if(blob != null) return "https://storage.cloud.google.com/" + bucket + "/" + code + "/" + sysName;
             return "fail";
 
         } catch (Exception e) {
@@ -38,17 +43,30 @@ public class GCSFile {
     }
 
     /** GCS 파일 서버에 파일 및 이미지 삭제 **/
-    public String deleteFileGcs (String sysName) {
+    public String deleteFileGcs(String sysName, String code) {
+        BlobId blobId = BlobId.of(bucket, code + "/" + sysName);
         try {
-            BlobId blobId = BlobId.of(bucket, sysName);
-            boolean result = storage.delete(blobId);
-            if(result) return "ok";
-            return "fail";
+            // 파일 객체를 가져옴
+            Blob blob = storage.get(blobId);
 
+            if (blob == null) {
+                logger.warn("파일이 존재하지 않습니다: " + blobId.toString());
+                return "fail";
+            }
+            
+            boolean result = storage.delete(blobId);
+
+            if (result) {
+                return "ok";
+            } else {
+                logger.warn("파일 삭제 실패: 파일이 존재하지 않거나 다른 이유로 실패, BlobId: " + blobId.toString());
+                return "fail";
+            }
         } catch (Exception e) {
-            e.printStackTrace();
+            logger.error("파일 삭제 실패: " + sysName + ", 코드: " + code, e);
             return "fail";
         }
     }
+
 
 }
